+++ HISTORY.md
@@ -1,6 +1,12 @@
 Crass Change History
 ====================
 
+git
+---
+
+* Fixed: `:at_rule` nodes now have a `:name` key.
+
+
 0.0.1 (2013-09-27)
 ------------------